refactor: move split shard command under shard
