+++ cmd/admin/shard/cmd.go
@@ -1,0 +1,31 @@
+// Copyright 2023-2026 The Oxia Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package shard
+
+import (
+	splitcmd "github.com/oxia-db/oxia/cmd/admin/shard/split"
+
+	"github.com/spf13/cobra"
+)
+
+var Cmd = &cobra.Command{
+	Use:   "shard",
+	Short: "Oxia admin operations on shards",
+	Long:  `Oxia admin operations on shards`,
+}
+
+func init() {
+	Cmd.AddCommand(splitcmd.Cmd)
+}
